Create ListOperationInPython.py
A list is like a dynamic and heterogeneous array. Arrays are used in languages like C. Unlike arrays, lists are dynamically sized and can store data of different types. It is one of the sequential data types in Python.

Important points about Lists:

Lists are mutable, meaning that after creation, we can modify their elements.
It consists of data separated by commas inside square braces-[].
Lists allow duplicate elements.
We use indexing to access the list elements (0 to length - 1), and negative indexing is allowed too (-1 to -length).
We can nest another data structure as an element inside a list. Nesting another list makes it a multi-dimensional list.
Using Slicing, we can get sub-lists of a list for which we need to use the slice operator - [:].
List comprehension is used to create new lists from existing lists or other data structures. It is faster than creating a list using the regular for loop.
We can use functions like max(), min() and sum() to find the maximum, minimum and sum of the elements in the list.
We can concatenate lists using + and repeat elements using * operators.
To take a list as an input, we use input() and then the split(separator) to convert the string into a list.